refactor(F719): merge skill-parse.ts into skill-meta.ts
skill-parse.ts and skill-meta.ts both had parseManifestSkillMeta and
SkillMeta — duplicate implementations from separate extraction passes.

Consolidate: add requiresMcp + resolveSkillMcpStatuses to skill-meta.ts,
replace listSkillDirs with listSourceSkillNames (from skill-source.ts),
delete skill-parse.ts (-112 lines).
